[MOD] EventDetailViewModel 수정
